@@ -13,13 +14,13 @@
     "Dário Vieira <dario.junior3@gmail.com>"
   ],
   "engines": {
-    "node": "=14.19.3"
+    "node": "=19.1.0"
   },
   "license": "AGPL-3.0",
   "preferGlobal": true,
   "private": false,
   "bin": {
-    "gbot": "./boot.js"
+    "gbot": "./boot.cjs"
   },
   "repository": {
     "type": "git",
@@ -138,50 +139,50 @@
     "rimraf": "3.0.2",
     "safe-buffer": "5.2.1",
     "scanf": "1.1.2",
-    "sequelize": "6.5.0",
-    "sequelize-cli": "6.2.0",
-    "sequelize-typescript": "2.1.0",
-    "simple-git": "2.39.1",
+    "sequelize": "6.25.6",
+    "sequelize-cli": "6.5.2",
+    "sequelize-typescript": "2.1.5",
+    "simple-git": "3.15.0",
     "speakingurl": "14.0.1",
-    "sppull": "2.7.0",
+    "sppull": "3.0.1",
     "ssr-for-bots": "1.0.1-c",
     "strict-password-generator": "1.1.2",
     "swagger-client": "2.1.18",
-    "tabulator-tables": "^5.2.6",
-    "tedious": "14.0.0",
+    "tabulator-tables": "5.4.2",
+    "tedious": "15.1.2",
     "textract": "2.5.0",
-    "twitter-api-v2": "1.12.7",
-    "typescript": "3.6.4",
-    "typescript-rest-rpc": "^1.0.10",
-    "url-join": "4.0.1",
+    "twitter-api-v2": "1.12.9",
+    "typescript": "4.9.3",
+    "typescript-rest-rpc": "1.0.10",
+    "url-join": "5.0.0",
     "vbscript-to-typescript": "1.0.8",
-    "vhost": "^3.0.2",
-    "vm2": "^3.9.11",
-    "vm2-process": "^2.1.1",
+    "vhost": "3.0.2",
+    "vm2": "3.9.11",
+    "vm2-process": "2.1.1",
     "walk-promise": "0.2.0",
     "washyourmouthoutwithsoap": "1.0.2",
-    "whatsapp-web.js": "1.18.0",
-    "winston": "^2.4.2",
-    "winston-logs-display": "^1.0.0",
-    "yarn": "^1.22.19"
+    "whatsapp-web.js": "1.18.3",
+    "winston": "3.8.2",
+    "winston-logs-display": "1.0.0",
+    "yarn": "1.22.19"
   },
   "devDependencies": {
-    "@types/puppeteer": "5.4.6",
-    "@types/url-join": "4.0.0",
+    "@types/puppeteer": "5.4.7",
+    "@types/url-join": "4.0.1",
     "@types/winston": "2.4.4",
-    "ban-sensitive-files": "1.9.15",
-    "commitizen": "4.2.4",
+    "ban-sensitive-files": "1.9.18",
+    "commitizen": "4.2.5",
     "cz-conventional-changelog": "3.3.0",
     "dependency-check": "4.1.0",
     "git-issues": "1.3.1",
     "license-checker": "25.0.1",
-    "ngrok": "4.0.1",
+    "ngrok": "4.3.3",
     "nsp": "3.2.1",
     "prettier-standard": "16.4.1",
-    "semantic-release": "17.4.3",
-    "simple-commit-message": "4.1.2",
+    "semantic-release": "19.0.5",
+    "simple-commit-message": "4.1.3",
     "travis-deploy-once": "5.0.11",
-    "ts-node": "10.0.0",
+    "ts-node": "10.9.1",
     "tslint": "6.1.2"
   },
   "eslintConfig": {
