@@ -38,11 +38,11 @@
 
 import { GBMinInstance, IGBCoreService, IGBInstance } from 'botlib';
 import { CreateOptions } from 'sequelize/types';
-const fs = require('fs');
-const urlJoin = require('url-join');
-import { GBServer } from '../../../src/app';
-import { GuaribasInstance } from '../models/GBModel';
-import { GBConfigService } from './GBConfigService';
+import Fs from 'fs';
+import urlJoin from 'url-join';
+import { GBServer } from '../../../src/app.js';
+import { GuaribasInstance } from '../models/GBModel.js';
+import { GBConfigService } from './GBConfigService.js';
 
 /**
  * Handles the importing of packages.
@@ -56,7 +56,7 @@ export class GBImporter {
 
   public async importIfNotExistsBotPackage(botId: string,
                                            packageName: string, localPath: string, additionalInstance: IGBInstance = null) {
-    const settingsJson = JSON.parse(fs.readFileSync(urlJoin(localPath, 'settings.json'), 'utf8'));
+    const settingsJson = JSON.parse(Fs.readFileSync(urlJoin(localPath, 'settings.json'), 'utf8'));
     if (botId === undefined) {
       botId = settingsJson.botId;
     }
@@ -107,8 +107,8 @@ export class GBImporter {
     localPath: string,
     settingsJson: any
   ) {
-    const packageJson = JSON.parse(fs.readFileSync(urlJoin(localPath, 'package.json'), 'utf8'));
-    const servicesJson = JSON.parse(fs.readFileSync(urlJoin(localPath, 'services.json'), 'utf8'));
+    const packageJson = JSON.parse(Fs.readFileSync(urlJoin(localPath, 'package.json'), 'utf8'));
+    const servicesJson = JSON.parse(Fs.readFileSync(urlJoin(localPath, 'services.json'), 'utf8'));
 
     const fullSettingsJson = { ...GBServer.globals.bootInstance, ...packageJson, ...settingsJson, ...servicesJson };
 
