@@ -30,11 +30,9 @@
 |                                                                             |
 \*****************************************************************************/
 
-const urlJoin = require('url-join');
 import { GBService } from 'botlib';
-const fs = require('fs');
-var AdmZip = require("adm-zip");
-
+import Fs from 'fs';
+import AdmZip from 'adm-zip';
 
 /**
  * Support for Whatsapp.
@@ -51,7 +49,7 @@ export class TeamsService extends GBService {
   }
 
   public async getManifest(marketplaceId, botName, botDescription, id, packageName, yourName) {
-    let content = fs.readFileSync('teams-manifest.json', 'utf8');
+    let content = Fs.readFileSync('teams-manifest.json', 'utf8');
 
     content = content.replace(/\@\@marketplaceId/gi, marketplaceId);
     content = content.replace(/\@\@botName/gi, botName);
